1. Add VDM flags; 2. Under api>public, transfer cmis related functions from c_cmis.py  to cmis.py
