+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-addopts = --cov=sonic_platform_base tests/sonic_xcvr/test_ccmis.py --cov-report html --cov-report term --cov-report xml --junitxml=test-results.xml -v
+addopts = --cov=sonic_platform_base tests/sonic_xcvr/test_cmis.py --cov-report html --cov-report term --cov-report xml --junitxml=test-results.xml -v
