+++ sonic_platform_base/sonic_xcvr/xcvr_api_factory.py
@@ -9,7 +9,7 @@ from .xcvr_eeprom import XcvrEeprom
 # TODO: remove the following imports
 from .codes.public.sff8024 import Sff8024
 from .codes.public.cmis_code import CmisCode
-from .api.public.c_cmis import CCmisApi
+from .api.public.cmis import CmisApi
 from .mem_maps.public.cmis import CmisMemMap
 
 class XcvrApiFactory(object):
@@ -26,4 +26,4 @@ class XcvrApiFactory(object):
         codes = {'sff8024':Sff8024, 'cmis_code':CmisCode}
         mem_map = CmisMemMap(codes)
         xcvr_eeprom = XcvrEeprom(self.reader, self.writer, mem_map)
-        return CCmisApi(xcvr_eeprom)
+        return CmisApi(xcvr_eeprom)
